+++ example_global_mode/android/app/src/main/kotlin/com/example/example_global_mode/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.example_global_mode
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}